First round of file renames in cli
