+++ packages/cli/src/commands/publish/PublishCliRemove.ts
